refs #15793_15794 refactored code, removed blocking trigger check on method
