+++ app/models/hbx_enrollment.rb
@@ -636,11 +636,11 @@ class HbxEnrollment
   def mid_year_plan_change_notice
     if self.census_employee.present?
       begin
-        if (self.enrollment_kind != "open_enrollment" || self.census_employee.new_hire_enrollment_period.present?) && benefit_group.plan_year.open_enrollment_end_on > TimeKeeper.date_of_record
+        if (self.enrollment_kind != "open_enrollment" || self.census_employee.new_hire_enrollment_period.present?)
           if self.benefit_group.is_congress
-            ShopNoticesNotifierJob.perform_later(self.census_employee.id.to_s, "ee_mid_year_plan_change_notice_congressional")
+            ShopNoticesNotifierJob.perform_later(self.employer_profile.id.to_s, "ee_mid_year_plan_change_congressional_notice", hbx_enrollment: self.hbx_id.to_s)
           else
-            ShopNoticesNotifierJob.perform_later(self.census_employee.id.to_s, "employee_mid_year_plan_change_non_congressional")
+            ShopNoticesNotifierJob.perform_later(self.employer_profile.id.to_s, "ee_mid_year_plan_change_non_congressional_notice", hbx_enrollment: self.hbx_id.to_s)
           end
         end  
       rescue Exception => e
