+++ app/views/notices/shop_employer_notices/ee_mid_year_plan_change_notice_congressional.html.erb
@@ -3,9 +3,6 @@
 <%= render partial: "notices/shared/shop_address", :locals => {notice: notice} %>
 <%= render partial: "notices/shared/date", :locals => {notice: notice} %>
 
-<br><br>
-<%= render partial: 'notices/shared/date', :locals => {notice: notice} %>
-
 <table class="total_table" border="0">
   <tbody>
   <tr>
@@ -17,18 +14,16 @@
 <table class="total_table" border="0">
   <tbody>
   <tr>
-    <td>Dear <%= notice.primary_fullname %>:</td>
+    <td>Dear <%= notice.employer_name %>:</td>
   </tr>
   </tbody>
 </table>
 
 <%= render partial: 'notices/shared/paragraph', :locals => {
-content: "This notice is to inform you that an employee of #{notice.employer_name}, #{notice.employee_fullname}, has made a change to their employer-sponsored coverage. The change will take effect on #{notice.sep.effective_on}. You can review the details of this change in your employer account. This change will be reflected on your next invoice."} %>
-
-
+  content: "This notice is to inform you that an employee of #{notice.employer_name}, #{notice.employee.primary_fullname}, has made a change to their employer-sponsored coverage. The change will take effect on #{notice.enrollment.effective_on}. You can review the details of this change in your employer account. This change will be reflected on your next invoice."} %>
+<br>
 <%= render partial: 'notices/shared/paragraph_with_headers', :locals => {
-contents: [""],
-heading: 'Why was the Change Made?', sub_heading: ''} %>
+  contents: [""], heading: 'Why was the Change Made?', sub_heading: ''} %>
 
 <%= render partial: 'notices/shared/unordered_list', locals: {
   previous: "",
@@ -36,32 +31,18 @@ heading: 'Why was the Change Made?', sub_heading: ''} %>
    ]} %>
 
 <%= render partial: 'notices/shared/paragraph_with_headers', :locals => {
-  contents: ["Please contact your benefits office for further assistance.  You can also contact #{Settings.site.short_name} with any questions:"],
-  heading: 'For Questions or Assistance:', sub_heading: ''} %>
-<table class="total_table" border="0">
-  <tbody>
-  <tr>
-    <td>
-      <ul style="list-style: circle solid;">
-        <li style="margin-top: 5px;">
-          By calling <%=Settings.contact_center.phone_number%>
-        </li>
-        <li style="margin-top: 5px;">
-          TTY: 711
-        </li>
-        <li style="margin-top: 5px;">
-          Online at <a href="www.DCHealthLink.com">www.DCHealthLink.com</a>
-        </li>
-      </ul>
-    </td>
-  </tr>
-  </tbody>
-</table>
+  contents: ["Please contact your benefits office for further assistance.  You can also contact #{Settings.site.short_name} with any questions:"], heading: 'For Questions or Assistance:', sub_heading: ''} %>
+
+<%= render partial: 'notices/shared/unordered_list', :locals => { listings:
+    ["By calling #{notice.hbe.phone}",
+    "TTY: #{Settings.contact_center.tty}",
+    "Online at <a href=#{Settings.site.home_url}>#{Settings.site.website_name}</a>"]} %>
+
 <% if notice.broker.present? %>
 <%= render partial: 'notices/shared/broker_contact_info', :locals => {broker: notice.broker} %>
 <% else %>
 <%= render partial: 'notices/shared/paragraph', :locals => {
-  content: "If you do not currently have a broker, you can reach out to one of our many trained experts by clicking on the “Find a Broker” link in your employer account or calling #{Settings.contact_center.phone_number}."} %>
+  content: "If you do not currently have a broker, you can reach out to one of our many trained experts by clicking on the “Find a Broker” link in your employer account or calling #{notice.hbe.phone}."} %>
 <% end %>
 
 <%= render partial: 'notices/shared/paragraph', :locals => {
