+++ app/views/notices/shop_employer_notices/employee_mid_year_plan_change_non_congressional.html.erb
@@ -2,6 +2,7 @@
 <%= render partial: 'notices/shared/logo_and_address_shop' %>
 <%= render partial: "notices/shared/shop_address", :locals => {notice: notice} %>
 <%= render partial: "notices/shared/date", :locals => {notice: notice} %>
+
 <table class="total_table" border="0">
   <tbody>
   <tr>
@@ -13,49 +14,35 @@
 <table class="total_table" border="0">
   <tbody>
   <tr>
-    <td>Dear <%= notice.primary_fullname %>:</td>
+    <td>Dear <%= notice.employer_name %>:</td>
   </tr>
   </tbody>
 </table>
 
 <%= render partial: 'notices/shared/paragraph', :locals => {
-  content: "This notice is to inform you that an employee of #{notice.employer_name}, #{notice.primary_fullname}, has made a change to their employer-sponsored coverage. The change will take effect on #{notice.sep.effective_on}. You can review the details of this change in your employer account. This change will be reflected on your next invoice."} %>
-
+  content: "This notice is to inform you that an employee of #{notice.employer_name}, #{notice.employee.primary_fullname}, has made a change to their employer-sponsored coverage. The change will take effect on #{notice.enrollment.effective_on}. You can review the details of this change in your employer account. This change will be reflected on your next invoice."} %>
+<br>
 <%= render partial: 'notices/shared/paragraph_with_headers', :locals => {
-    contents: [""], heading: 'Why was the Change Made?', sub_heading: ''} %>
+  contents: [""], heading: 'Why was the Change Made?', sub_heading: ''} %>
 
 <%= render partial: 'notices/shared/unordered_list', locals: {
   previous: "",
   listings: ["<strong>New Hire Open Enrollment:</strong> A new hire may have chosen to enroll in employer-sponsored health insurance. Once added to the employer’s roster, a new hire’s 30-day new hire open enrollment period is triggered. Newly-hired employees have 30 days from the date they are added to the roster OR their date of hire, whichever is later, to enroll in employer-sponsored insurance or waive coverage.","<strong>Qualifying Event:</strong> Your employee may have experienced a qualifying event that made them eligible to make changes to their enrollment. Employees who experience a qualifying event during the plan year may be eligible for a Special Enrollment Period (SEP) which allows them to enroll in a plan, make changes to their plan selection, add or drop family members, or drop coverage altogether.", "<strong>Dependent Aged Off of Coverage:</strong> Your employee may have had a dependent on their policy that is no longer eligible for coverage. Dependent children are eligible to remain enrolled on their parents’ plan up to age 26. Their coverage extends through the end of the calendar year during which they turn 26 and automatically terminates on December 31<sup>st</sup> of that year."]} %>
 
-<%= render partial: 'notices/shared/paragraph_with_headers', :locals => {
-                                                                   contents: ["Please contact your broker for further assistance.  You can also contact #{Settings.site.short_name} with any questions:"],
-                                                                   heading: 'For Questions or Assistance:', sub_heading: ''} %>
-<table class="total_table" border="0">
-  <tbody>
-  <tr>
-    <td>
-      <ul style="list-style: circle solid;">
-        <li style="margin-top: 5px;">
-          By calling <%=Settings.contact_center.phone_number%>
-        </li>
-        <li style="margin-top: 5px;">
-          TTY: 711
-        </li>
-        <li style="margin-top: 5px;">
-          Online at <a href="www.DCHealthLink.com">www.DCHealthLink.com</a>
-        </li>
-      </ul>
-    </td>
-  </tr>
-  </tbody>
-</table>
+<%= render partial: 'notices/shared/paragraph_with_headers', :locals => { 
+  contents: ["Please contact your broker for further assistance.  You can also contact #{Settings.site.short_name} with any questions:"], heading: 'For Questions or Assistance:', sub_heading: '', top_margin: '10px'} %>
+
+<%= render partial: 'notices/shared/unordered_list', :locals => { listings:
+    ["By calling #{notice.hbe.phone}",
+    "TTY: #{Settings.contact_center.tty}",
+    "Online at <a href=#{Settings.site.home_url}>#{Settings.site.website_name}</a>"]} %>
+
 <% if notice.broker.present? %>
-    <%= render partial: 'notices/shared/broker_contact_info', :locals => {broker: notice.broker} %>
+  <%= render partial: 'notices/shared/broker_contact_info', :locals => {broker: notice.broker} %>
 <% else %>
   <%= render partial: 'notices/shared/paragraph', :locals => {
-  content: "If you do not currently have a broker, you can reach out to one of our many trained experts by clicking on the “Find a Broker” link in your employer account or calling #{Settings.contact_center.phone_number}."} %>
+  content: "If you do not currently have a broker, you can reach out to one of our many trained experts by clicking on the “Find a Broker” link in your employer account or calling #{notice.hbe.phone}."} %>
 <% end %>
 
 <%= render partial: 'notices/shared/paragraph', :locals => {
-                                                content: "This notice is being provided in accordance with 45 CFR 155.720(h)"} %>
+  content: "This notice is being provided in accordance with 45 CFR 155.720(h)"} %>
