+++ db/seedfiles/notice_triggers_seed.rb
@@ -613,27 +613,6 @@ shop_notice_triggers = [
     ]
   },
   {
-    hbx_id: 'SHOP40',
-    title: 'Employee_Mid-Year Plan Change-Non-Congressional',
-    description: 'Employee mid year plan change when an eligiblity determination is reached',
-    resource_name: 'employer',
-    event_name: 'employee_mid_year_plan_change_non_congressional',
-    notice_triggers: [
-      {
-        name: 'Employee Made Mid-Year Plan',
-        notice_template: 'notices/shop_employer_notices/employee_mid_year_plan_change_non_congressional',
-        notice_builder: 'ShopEmployerNotices::EmployeeMidYearPlanChangeNonCongressional',
-        mpi_indicator: 'MPI_SHOP40',
-        notice_trigger_element_group: {
-          market_places: ['shop'],
-          primary_recipients: ["employer"],
-          primary_recipient_delivery_method: ["secure_message"],
-          secondary_recipients: []
-        }
-      }
-    ]
-  },
-  {
     hbx_id: 'SHOP35',
     title: 'Special Enrollment Period Denial',
     description: 'EE SEP Requested by Employee outside of allowable time frame',
@@ -676,17 +655,17 @@ shop_notice_triggers = [
     ]
   },
   {
-    hbx_id: 'SHOP37',
-    title: 'Employee mid-year plan change',
+    hbx_id: 'SHOP_D037',
+    title: 'Employee has made a change to their employer-sponsored coverage selection',
     description: 'Employee mid year plan change when an eligiblity determination is reached',
     resource_name: 'employer',
-    event_name: 'ee_mid_year_plan_change_notice_congressional',
+    event_name: 'ee_mid_year_plan_change_congressional_notice',
     notice_triggers: [
       {
-        name: 'Employee mid year plan change notice',
+        name: 'Employee Mid-Year Plan change Congressional',
         notice_template: 'notices/shop_employer_notices/ee_mid_year_plan_change_notice_congressional',
-        notice_builder: 'ShopEmployerNotices::EeMidYearPlanChangeNoticeCongressional',
-        mpi_indicator: 'MPI_SHOP37',
+        notice_builder: 'ShopEmployerNotices::EeMidYearPlanChangeNotice',
+        mpi_indicator: 'SHOP_D037',
         notice_trigger_element_group: {
           market_places: ['shop'],
           primary_recipients: ["employer"],
@@ -697,17 +676,17 @@ shop_notice_triggers = [
     ]
   },
   {
-    hbx_id: 'SHOP40',
-    title: 'Employee Mid-Year Plan Change',
+    hbx_id: 'SHOP_D040',
+    title: 'Employee has made a change to their employer-sponsored coverage selection',
     description: 'Employee mid year plan change when an eligiblity determination is reached',
     resource_name: 'employer',
-    event_name: 'employee_mid_year_plan_change_non_congressional',
+    event_name: 'ee_mid_year_plan_change_non_congressional_notice',
     notice_triggers: [
       {
-        name: 'Employee Mid-Year Plan change',
+        name: 'Employee Mid-Year Plan change Non-Congressional',
         notice_template: 'notices/shop_employer_notices/employee_mid_year_plan_change_non_congressional',
-        notice_builder: 'ShopEmployerNotices::EmployeeMidYearPlanChangeNonCongressional',
-        mpi_indicator: 'MPI_SHOP40',
+        notice_builder: 'ShopEmployerNotices::EeMidYearPlanChangeNotice',
+        mpi_indicator: 'SHOP_D040',
         notice_trigger_element_group: {
           market_places: ['shop'],
           primary_recipients: ["employer"],
